k8s/apps/monitoring: refactor solana monitoring for Jump nodes
We need to run one solana_exporter instance per Solana node
for now (further improvements to this are tracked in T929).

Our abstraction layer kind of starts to break down
with customer services and RPC-only nodes - we should
eventually add a `customer` tag and get service tags back
in line with the DNS names for Terra and NEAR (T828).

We'll get duplicate notifications for validator downtime,
but alertmanager grouping will take care of it.

Now that we have multiple nodes per network, we can do the
same thing as for Tendermint and check delta to cluster
via the new SolanaBehindCluster alert.

Test Plan: Tested very carefully on pa-dev

After deployment, manually clean up the old statefulsets:

    kc delete statefulsets.apps solana-exporter-tds-0 solana-exporter-solana-0

GitOrigin-RevId: efb3a54321f2f30baa1a8deb1b653c862b29ddf6