k8s/apps/certmanager: add global ClusterIssuer
By default, certificates can only reference issuers in the same namespace,
except for ClusterIssuer, which can be referenced globally. We don't want to
expose the CloudFlare API key to every namespace that needs an individual
certificate, or create separate ACME accounts for each.

However, this comes with a downside: cert-manager has no allowlist and will
happily issue certificate using your issuer for every namespace. This is fine
for our single-tenant cluster which is only written to by our own deployment,
but if we ever end up with less trusted namespaces we can probably work around
this by limiting the namespaces the controller has read access to.

(a similar problem exists for wildcard certificates: if we want to use them
from multiple namespaces, we'll have to resort to...copying:
https://cert-manager.io/docs/faq/kubed/)

Test Plan: Deployed the thing, it dutifully created an account key.

X-Origin-Diff: phab/D725
GitOrigin-RevId: a27566e847dbe5fe950e440bbc6d08945032a23d