jenkins: force upgrade of plugins
This forces upgrade of all plugins always by the Jenkins controller
image entrypoint machinery, which is needed in case of plugins that have
been manually installed or came stock with the image. This shouldn't be
an issue in practice, but might happen when dealing with dev/test
Jenkins deployments (as it has now).

X-Origin-Diff: phab/D780
GitOrigin-RevId: e4b661c43ac884658d79d8675693981b4620d374