jenkins: generate sum fields for secrets
In preparation for moving the controller configmaps to secrets, we
extend the 'sum' logic previous implemented to also work for secrets.

This _should_ be safe for all secrets everywhere, but does technically
act as an oracle under some scenarios. However, for now, we only use
this for the Jenkins deployment, in which the secrets are known to not
be susceptible to any attacks related to their cryptographic sum being
available.

X-Origin-Diff: phab/D782
GitOrigin-RevId: 7407c0d017f0f4feb1d23c3867f3df7abba54840