jenkins: configure nodes/agents in controller
This implements a fairly nasty hack that allows us to preconfigure (and
effectively have access to in CUE) agent connection secrets.

This unfortunately can only be done by controlling the HMAC key used to
create these secrets. In an ideal world, each agent would have an
individually configurable secret via CasC, but that is unfortunately not
the case.

Here, something like a proper operator would help us, as it could
sequence the bringup of the controller, retrieval of the generated
secret and then the bringup of agents. If we keep hitting issues like
this, it might be worth it to migrate over, as the complexity of an
operator might be better than the complexity of similar hacks.

X-Origin-Diff: phab/D785
GitOrigin-RevId: 745476f6a2a7bcddd53a0569c4908e962280a837