jenkins: set plugin mode to 0644
This hopefully fixes plugin updates - the entrypoint machinery from the
Jenkins controller image attempts to copy them over, but seems to fail
silently if it cannot write the target plugin.

As the plugin files are copied verbatim and not `install`ed, whatever
mode they are in /usr/share/jenkins/ref/plugins/ they will end up with
the same in /var/jenkins_home/plugins. Thus, if that mode is 0555,
updates will fail.

This naturally is a bug in the Jenkins container entrypoint code, but we
attempt to work around it here by setting the plugin files mode to
something writable.

I was not able to verify whether this works, as by the point I attempted
this fix I've inadvertendly mangled up jenkins-dev's
/var/jenkins_home/plugins directory by a spurious unzip, which I then
proceeded to just nuke. We might thus have to revisit this on another
plugin update, and looking at how much Jenkins plugins get updated, this
will be soon enough that I won't go out of my way to attempt to
reproduce this and see if this fix actually works.

X-Origin-Diff: phab/D792
GitOrigin-RevId: d1b3eb42fc4cc6b0dfb84c86b7b13babd71ac9ad