k8s/apps/gerrit: bring our own oauth.jar
This will allow us to customize it at some point (we'll want to deny
*.monogon.tech logins via GitHub), and it's one less thing we pull
from that creepy public Jenkins instance :-)

Bazlets are on the 3.5 API version, but that seems to be fine. We would
have to build inside the Gerrit tree otherwise, which is not something
we can easily do within this WORKSPACE.

Tested and works on gerrit-dev.

Change-Id: I08cf2f8ec1509a3fb934c625881c3caafb97e9fe
Reviewed-on: https://review.monogon.dev/c/infralib/+/1119
Reviewed-by: Serge Bazanski <serge@monogon.tech>